reads config.ini or creates one if it doesn't exist

+++ DroneClient/Config.cs
@@ -9,21 +9,35 @@ namespace DroneClient
 		
 		public Config ()
 		{
-            try {
-                string[] lines = File.ReadAllLines(DCConstants.ConfigFile);
-                foreach (string line in lines) {
-                string[] data = line.Split(':');
-                if (data[0].Equals("username")) DCConstants.Username = data[1];
-                else if (data[0].Equals("password")) DCConstants.Password = data[1];
-                }
-            }
-            catch {
-				File.Create(DCConstants.ConfigFile);
-				StreamWriter writer = new StreamWriter(DCConstants.ConfigFile);
-				writer.WriteLine("username:");
-				writer.WriteLine("password:");
-				writer.Close();
-            }
+			if (File.Exists (DCConstants.ConfigFile)) {
+				string[] lines;
+
+				try {
+					lines = File.ReadAllLines (DCConstants.ConfigFile);
+					if (lines != null) {
+						foreach (string line in lines) {
+							string[] data = line.Split (':');
+							if (data.Length >= 2) {
+								if (data [0].Equals ("username")) DCConstants.Username = data [1];
+								else if (data [0].Equals ("password")) DCConstants.Password = data [1];
+							}
+						}
+					}
+				} catch {
+				}
+
+
+			} else {
+				try {
+					//File.Create(DCConstants.ConfigFile);
+					StreamWriter writer = new StreamWriter(DCConstants.ConfigFile);
+					writer.WriteLine("username:");
+					writer.WriteLine("password:");
+					writer.Close();
+				} catch {
+				}
+			}
+            
         }
 		
 		public static void checkcreds ()
